solar-cleaning/backend/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.35
python-dotenv==0.19.2
Flask-CORS==3.0.10
Flask-Login==0.6.2
Werkzeug==2.0.3
